Add skeleton/rough outline of service flow.
The base functionality is there to call the Toggl API and act according
to what was presented. Notifications are sent to my iPhone if nothing
is running or if something has been running for more than a threshold.

Many things are duplicated and hardcoded at the moment. That's fine for
now.